diskutil initial disassembly
I think the program and data have been largely sorted out.  Might even be able to change it and still have it work.
Fair number of the simpler leaf subroutines identified.  Got a bit lazy on parts of the console output which is very similar to z80ctl, but at least all the is in a contiguous section and most of the console routines used by the main code are named.